+++ CHANGELOG.md
@@ -6,5 +6,6 @@
   - [slimScroll](https://github.com/rochal/jQuery-slimScroll)
   - [iCheck](https://github.com/fronteed/iCheck)
 - FIX:
-  - Install Generator: Add stylesheet_engine when injecting file
+  - Install Generator: Use stylesheet_engine when injecting data in file
+  - Plugin Generator: Use stylesheet_engine when injecting data in file